New Check Quantity tests and updated configs.

See merge request lego-front/auto-tests!204

+++ src/test/java/com/leroy/magmobile/api/data/sales/orders/ResOrderProductCheckQuantityData.java
@@ -9,7 +9,7 @@ public class ResOrderProductCheckQuantityData {
     private String lineId;
     private String lmCode;
     private Double quantity;
-    private Integer availableStock;
+    private Double availableStock;
     private String title;
     private String barCode;
     private String departmentId;
